+++ server/src/main/java/de/njsm/stocks/server/v2/db/FoodHandler.java
@@ -97,7 +97,7 @@ public class FoodHandler extends CrudDatabaseHandler<FoodRecord, Food> {
                             .and(getVersionField().eq(item.version)
                                     .and(FOOD.NAME.ne(newName)
                                             .or(FOOD.EXPIRATION_OFFSET.ne(expirationOffset))
-                                            .or(FOOD.LOCATION.ne(location)))
+                                            .or(FOOD.LOCATION.isDistinctFrom(location)))
                             )
             )
                     .map(this::notFoundMeansInvalidVersion);